Java JNI functions for local playback

+++ Library/TeamTalkLib/avstream/FFMpeg3Streamer.h
@@ -30,6 +30,8 @@
 
 #include "MediaStreamer.h"
 
+#include <stdint.h>
+
 // compatible with ffmpeg tag n1.1.13
 
 void InitAVConv();
@@ -56,16 +58,16 @@ protected:
 private:
     void Run();
 
-    int ProcessAudioBuffer(struct AVFilterContext* aud_buffersink_ctx,
-                           struct AVFrame* filt_frame,
-                           struct AVStream* aud_stream,
-                           ACE_UINT32 start_time,
-                           ACE_UINT32& start_offset);
-    int ProcessVideoBuffer(struct AVFilterContext* vid_buffersink_ctx,
-                           struct AVFrame* filt_frame,
-                           struct AVStream* vid_stream,
-                           ACE_UINT32 start_time,
-                           ACE_UINT32& start_offset);
+    int64_t ProcessAudioBuffer(struct AVFilterContext* aud_buffersink_ctx,
+                               struct AVFrame* filt_frame,
+                               struct AVStream* aud_stream,
+                               ACE_UINT32 start_time,
+                               ACE_UINT32& start_offset);
+    int64_t ProcessVideoBuffer(struct AVFilterContext* vid_buffersink_ctx,
+                               struct AVFrame* filt_frame,
+                               struct AVStream* vid_stream,
+                               ACE_UINT32 start_time,
+                               ACE_UINT32& start_offset);
 };
 
 #endif
